feat: add clothing + error and type fix

+++ controllers/brandController.js
@@ -13,11 +13,14 @@ class BrandController {
     return res.json(brands);
   }
 
-  // доделать отлов ошибок, удаления того что уже удалено
-  async delete(req, res) {
+  async delete(req, res, next) {
     const { id } = req.params;
     if (!id) {
-      ApiError.badRequest();
+      return next(ApiError.badRequest("Не задан ID"));
+    }
+    const elem = await Type.findByPk(id);
+    if (!elem) {
+      return next(ApiError.badRequest("Нет такого элемента"));
     }
     try {
       await Brand.destroy({
@@ -27,7 +30,7 @@ class BrandController {
       });
       return res.status(200).json({ message: "Успешно удалено" });
     } catch (e) {
-      ApiError.internal(e);
+      next(ApiError.internal(e.message));
     }
   }
 }
